+++ packages/react-charts/src/components/ChartUtils/chart-helpers.ts
@@ -4,6 +4,9 @@ interface ChartClassNameInterface {
   className?: string;
 }
 
+// Copied from exenv
+export const canUseDOM = !!(typeof window !== 'undefined' && window.document && window.document.createElement);
+
 // Returns the class name that will be applied to the outer-most div rendered by the chart's container
 export const getClassName = ({ className }: ChartClassNameInterface) => {
   let cleanClassName;