+++ packages/react-charts/src/components/ChartUtils/index.ts
@@ -6,5 +6,6 @@ export * from './chart-label';
 export * from './chart-legend';
 export * from './chart-origin';
 export * from './chart-padding';
+export * from './chart-resize';
 export * from './chart-theme';
 export * from './chart-tooltip';