+++ packages/react-log-viewer/src/LogViewer/LogViewerSearch.tsx
@@ -1,15 +1,19 @@
-import React, { useEffect, useState } from 'react';
-import { NUMBER_INDEX_DELTA, DEFAULT_FOCUS, DEFAULT_INDEX } from './utils/constants';
+import React, { useContext, useEffect, useState } from 'react';
+import { NUMBER_INDEX_DELTA, DEFAULT_FOCUS, DEFAULT_INDEX, DEFAULT_SEARCH_INDEX } from './utils/constants';
 import { SearchInput, SearchInputProps } from '@patternfly/react-core';
-import { LogViewerToolbarContext } from './LogViewerContext';
+import { LogViewerToolbarContext, LogViewerContext } from './LogViewerContext';
+import { escapeString, searchForKeyword } from './utils/utils';
 
 export interface LogViewerSearchProps extends SearchInputProps {
   /** Place holder text inside of searchbar */
   placeholder: string;
+  /** Minimum number of characters required for searching */
+  minSearchChars: number;
 }
 
 export const LogViewerSearch: React.FunctionComponent<LogViewerSearchProps> = ({
   placeholder = 'Search',
+  minSearchChars = 1,
   ...props
 }) => {
   const [indexAdjuster, setIndexAdjuster] = useState(0);
@@ -20,10 +24,12 @@ export const LogViewerSearch: React.FunctionComponent<LogViewerSearchProps> = ({
     setCurrentSearchedItemCount,
     setRowInFocus,
     setSearchedWordIndexes,
-    rowInFocus,
     currentSearchedItemCount,
-    searchedInput
-  } = React.useContext(LogViewerToolbarContext);
+    searchedInput,
+    itemCount
+  } = useContext(LogViewerToolbarContext);
+
+  const { parsedData } = useContext(LogViewerContext);
 
   /* Defaulting the first focused row that contain searched keywords */
   useEffect(() => {
@@ -34,6 +40,26 @@ export const LogViewerSearch: React.FunctionComponent<LogViewerSearchProps> = ({
       setIndexAdjuster(0);
     }
   }, [searchedWordIndexes]);
+
+  /* Updating searchedResults context state given changes in searched input */
+  useEffect(() => {
+    let foundKeywordIndexes: (number | null)[] = [];
+    const adjustedSearchedInput = escapeString(searchedInput);
+
+    if (adjustedSearchedInput !== '' && adjustedSearchedInput.length >= minSearchChars) {
+      foundKeywordIndexes = searchForKeyword(adjustedSearchedInput, parsedData, itemCount || parsedData.length);
+
+      if (foundKeywordIndexes.length !== 0) {
+        setSearchedWordIndexes(foundKeywordIndexes);
+        scrollToRow(foundKeywordIndexes[DEFAULT_SEARCH_INDEX]);
+        setCurrentSearchedItemCount(DEFAULT_INDEX);
+      }
+    }
+
+    if (!adjustedSearchedInput) {
+      setRowInFocus(DEFAULT_FOCUS);
+    }
+  }, [searchedInput]);
 
   const hasFoundResults = searchedWordIndexes.length > 0 && searchedWordIndexes[0] !== -1;
 
@@ -47,30 +73,22 @@ export const LogViewerSearch: React.FunctionComponent<LogViewerSearchProps> = ({
 
   /* Moving focus over to next row containing searched word */
   const handleNextSearchItem = (): void => {
-    const oldIndex = searchedWordIndexes.indexOf(rowInFocus);
-    const adjustedSearchCount = currentSearchedItemCount + NUMBER_INDEX_DELTA;
-    let tempPosition = oldIndex;
-
-    if (oldIndex >= searchedWordIndexes.length - NUMBER_INDEX_DELTA) {
-      return null;
-    }
+    const adjustedSearchCount = (currentSearchedItemCount + NUMBER_INDEX_DELTA) % searchedWordIndexes.length;
 
     setCurrentSearchedItemCount(adjustedSearchCount);
-    scrollToRow(searchedWordIndexes[++tempPosition]);
+    scrollToRow(searchedWordIndexes[adjustedSearchCount]);
   };
 
   /* Moving focus over to next row containing searched word */
   const handlePrevSearchItem = (): void => {
-    const oldIndex = searchedWordIndexes.indexOf(rowInFocus);
-    const adjustedSearchCount = currentSearchedItemCount - NUMBER_INDEX_DELTA;
-    let tempPosition = oldIndex;
+    let adjustedSearchCount = currentSearchedItemCount - NUMBER_INDEX_DELTA;
 
-    if (oldIndex <= DEFAULT_INDEX) {
-      return null;
+    if (adjustedSearchCount < DEFAULT_INDEX) {
+      adjustedSearchCount += searchedWordIndexes.length;
     }
 
     setCurrentSearchedItemCount(adjustedSearchCount);
-    scrollToRow(searchedWordIndexes[--tempPosition]);
+    scrollToRow(searchedWordIndexes[adjustedSearchCount]);
   };
 
   return (